// src/gen/ride/wallet/v1alpha1/wallet_service.ts
export interface Transaction {
    /**
     * @generated from protobuf field: string transaction_id = 1;
     */
    transactionId: string;
    /**
     * @generated from protobuf field: string account_id = 2;
     */
    accountId: string;
    /**
     * @generated from protobuf field: int32 amount = 3;
     */
    amount: number;
    /**
     * @generated from protobuf field: ride.wallet.v1alpha1.TransactionType type = 4;
     */
    type: TransactionType;
    /**
     * @generated from protobuf field: string batch_id = 5;
     */
    batchId: string;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp create_time = 6;
     */
    createTime?: Timestamp;
}
/**
 * @generated from protobuf enum ride.wallet.v1alpha1.TransactionType
 */
export enum TransactionType {
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_DEBIT = 1;
     */
    DEBIT = 1,
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_CREDIT = 2;
     */
    CREDIT = 2
}
// ...
class Transaction$Type extends MessageType<Transaction> {
    constructor() {
        super("ride.wallet.v1alpha1.Transaction", [
            { no: 1, name: "transaction_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "account_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "amount", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 4, name: "type", kind: "enum", T: () => ["ride.wallet.v1alpha1.TransactionType", TransactionType, "TRANSACTION_TYPE_"] },
            { no: 5, name: "batch_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 6, name: "create_time", kind: "message", T: () => Timestamp }
        ]);
    }
    create(value?: PartialMessage<Transaction>): Transaction {
        const message = { transactionId: "", accountId: "", amount: 0, type: 0, batchId: "" };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<Transaction>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Transaction): Transaction {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string transaction_id */ 1:
                    message.transactionId = reader.string();
                    break;
                case /* string account_id */ 2:
                    message.accountId = reader.string();
                    break;
                case /* int32 amount */ 3:
                    message.amount = reader.int32();
                    break;
                case /* ride.wallet.v1alpha1.TransactionType type */ 4:
                    message.type = reader.int32();
                    break;
                case /* string batch_id */ 5:
                    message.batchId = reader.string();
                    break;
                case /* google.protobuf.Timestamp create_time */ 6:
                    message.createTime = Timestamp.internalBinaryRead(reader, reader.uint32(), options, message.createTime);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Transaction, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string transaction_id = 1; */
        if (message.transactionId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.transactionId);
        /* string account_id = 2; */
        if (message.accountId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.accountId);
        /* int32 amount = 3; */
        if (message.amount !== 0)
            writer.tag(3, WireType.Varint).int32(message.amount);
        /* ride.wallet.v1alpha1.TransactionType type = 4; */
        if (message.type !== 0)
            writer.tag(4, WireType.Varint).int32(message.type);
        /* string batch_id = 5; */
        if (message.batchId !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.batchId);
        /* google.protobuf.Timestamp create_time = 6; */
        if (message.createTime)
            Timestamp.internalBinaryWrite(message.createTime, writer.tag(6, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message ride.wallet.v1alpha1.Transaction
 */
export const Transaction = new Transaction$Type();